# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DraftCrane API
  description: |
    Backend API for DraftCrane (dc) -- a browser-based nonfiction book writing
    environment with AI assistance. Powered by Cloudflare Workers and Hono.

    ## Authentication

    Most endpoints require a valid **Clerk JWT** passed as a Bearer token in the
    `Authorization` header. Public endpoints (health check, Clerk webhook, OAuth
    callback) use their own verification mechanisms and are documented accordingly.

    ## Rate Limiting

    Standard endpoints enforce **120 requests per minute** per user (atomic
    time-window counters in D1). AI and export endpoints have tighter limits
    noted in their descriptions.

    ## Error Responses

    All errors follow a consistent shape with `error` (human-readable message),
    `code` (machine-readable enum), and `requestId` (ULID for tracing).
  version: 2.0.0
  contact:
    name: DraftCrane Engineering
  license:
    name: Proprietary
servers:
  - url: https://dc-api.automation-ab6.workers.dev
    description: Production
...
components:
  securitySchemes:
    clerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT token
  schemas:
    Error:
      type: object
      required:
        - error
        - code
        - requestId
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - AUTH_REQUIRED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - DRIVE_ERROR
            - DRIVE_NOT_CONNECTED
            - VALIDATION_ERROR
            - LAST_CHAPTER
            - AI_ERROR
            - EXPORT_ERROR
            - INTERNAL_ERROR
          description: Machine-readable error code
        requestId:
          type: string
          description: ULID request ID for tracing
...
    ResearchClip:
      type: object
      required:
        - id
        - projectId
        - sourceId
        - content
        - createdAt
      properties:
        id:
          type: string
        projectId:
          type: string
        sourceId:
          type: string
          nullable: true
        chapterId:
          type: string
          nullable: true
        content:
          type: string
        sourceTitle:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
            code: AUTH_REQUIRED
            requestId: 01HXYZ...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found
            code: NOT_FOUND
            requestId: 01HXYZ...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Title is required
            code: VALIDATION_ERROR
            requestId: 01HXYZ...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate limit exceeded
            code: RATE_LIMITED
            requestId: 01HXYZ...
...
security:
  - clerkBearerAuth: []
paths:
...
  /projects/{projectId}/research/sources/search:
    get:
      tags: [Research]
      summary: Full-text search across sources
      description: |
        Searches source content using FTS5 full-text search. Returns matching
        sources with relevant snippets.
      operationId: searchSources
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                required:
                  - results
                properties:
                  results:
                    type: array
                    items:
                      type: object
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{projectId}/research/clips:
    post:
      tags: [Research]
      summary: Save a research clip
      description: |
        Saves a text clip from a source or AI result. Optionally linked to
        a specific chapter.
      operationId: saveClip
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                sourceId:
                  type: string
                chapterId:
                  type: string
                sourceTitle:
                  type: string
      responses:
        '201':
          description: Clip saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchClip'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      tags: [Research]
      summary: List research clips
      description: |
        Returns all clips for a project. Optionally filter by chapter.
      operationId: listClips
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: chapterId
          in: query
          required: false
          schema:
            type: string
          description: Filter clips by chapter
      responses:
        '200':
          description: Clips list
          content:
            application/json:
              schema:
                type: object
                required:
                  - clips
                properties:
                  clips:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResearchClip'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /research/clips/{clipId}:
    delete:
      tags: [Research]
      summary: Delete a clip
      description: Hard-deletes a research clip.
      operationId: deleteClip
      parameters:
        - name: clipId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Clip deleted
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/research/query:
    post:
      tags: [Research]
      summary: AI research query (SSE streaming)
      description: |
        Sends a natural language query against the project's source materials.
        Retrieves relevant chunks via FTS5, then generates an AI response with
        source-attributed snippets. Streams via SSE. Rate limited to 10
        requests per minute.
      operationId: researchQuery
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Natural language research question
                instructionId:
                  type: string
                  description: Custom instruction ID to use
      responses:
        '200':
          description: SSE stream of research results
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{projectId}/research/analyze:
    post:
      tags: [Research]
      summary: AI source analysis (SSE streaming)
      description: |
        Analyzes a specific source document using AI. Streams the analysis
        via SSE. Rate limited to 10 requests per minute.
      operationId: analyzeSource
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sourceId
              properties:
                sourceId:
                  type: string
                instructionId:
                  type: string
                  description: Custom instruction ID to use
      responses:
        '200':
          description: SSE stream of analysis
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ─── AI ───────────────────────────────────────────────────────────────
...